requirements.txt
streamlit>=1.20.0
pymongo>=4.0
pandas
python-dotenv
